Fallback to safer error handling due to breaking changes in error messages in Go master
The exact error messages in #83 need to be identified, if they haven’t already been fixed in the Go standard library. However special casing this for Windows isn’t a priority. The log may get noisy for a while until I work out a better way to suppress the expected errors.